% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ad.R
\name{createAd}
\alias{createAd}
\title{createAd}
\usage{
createAd(body)
}
\arguments{
\item{body}{Body of request that includes ad settings. See examples 
and details.}
}
\description{
Create an ad, see details and example.
}
\author{
John Coene \email{john.coene@cmcm.com}
}

