% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_image.R
\name{createImage}
\alias{createImage}
\title{Upload an image}
\usage{
createImage(file)
}
\arguments{
\item{file}{Image data, supports JPEG, GIF and PNG.}
}
\description{
Upload an image
}
\details{
The list of valid values for the body can be found 
\href{here}{http://api.ori.cmcm.com/doc/#api-Campaign-create_campaign}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")

# upload image            
createImage(file = "path_to_R_version/library/pn/img/Rlogo.png")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}
}

