% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_objects.R
\name{createObject}
\alias{createObject}
\title{createObject}
\usage{
createObject(object, body)
}
\arguments{
\item{object}{Object to be created, see details for valid values.}

\item{body}{Body of request that includes campaigns settings. See examples.}
}
\description{
Create any objects, see details and example.
}
\details{
Valid values for \code{object}: 
\itemize{
\item \code{audience} see \code{\link{createAudience}} for details.
\item \code{campaign} see \code{\link{createCampaign}} for details.
\item \code{adset} see \code{\link{createAdset}} for details.
\item \code{ad} see \code{\link{createAd}} for details.
}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# set body of request for campaign
camp <- list(name = "test",
             budget_type = "daily",
             budget_daily = "1000", 
             pkg_name = "test",
             objective = "1", 
             app_type = "2",
             web_type = "1", 
             landing_page = "2")

# create campaign             
createObject(body = camp, object = "campaign")

# equivalent to
# createCampaign(body = camp)

# list campaigns
listCampaigns()
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

