% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_ad.R
\name{deleteAd}
\alias{deleteAd}
\title{deleteAd}
\usage{
deleteAd(ad.id)
}
\arguments{
\item{ad.id}{id of object to delete}
}
\description{
Delete ad.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# get ads
ads <- listAds()

# delete random adsets
deleteAd(ad.id = sample(ads$id, 1))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

