% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_campaign.R
\name{deleteCampaign}
\alias{deleteCampaign}
\title{Delete a campaign}
\usage{
deleteCampaign(campaign.id)
}
\arguments{
\item{campaign.id}{id of object to delete}
}
\description{
Delete a campaign.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# get campaigns
campaigns <- listCampaigns()

# delete random campaign
deleteCampaign(campaign.id = sample(campaigns$id, 1))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createCampaign}}
}

