% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_device.R
\name{dictDevice}
\alias{dictDevice}
\title{dictDevice}
\usage{
dictDevice()
}
\description{
fetch list of device models
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict mobile devices
head(dictDevice())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

