% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_interest.R
\name{dictInterest}
\alias{dictInterest}
\title{List available interests}
\usage{
dictInterest()
}
\description{
Fetch list of interests.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict interests
head(dictInterest())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

