% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_language.R
\name{dictLanguage}
\alias{dictLanguage}
\title{List valid languages}
\usage{
dictLanguage()
}
\description{
Fetch list of languages.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict languages
head(dictLanguage())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

