% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_state.R
\name{dictState}
\alias{dictState}
\title{dictState}
\usage{
dictState(country.code)
}
\arguments{
\item{country.code}{Code of country as returned by \code{\link{dictCountry}}, 
see examples.}
}
\description{
fetch list of States
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict countries
country <- dictCountry()

# dict states of random country
states <- dictState(country.mode = sample(country$code, 1))

# dict US states
us <- dictState(country.mode = "US")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

