% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esl-data.R
\docType{data}
\name{esl}
\alias{esl}
\title{ESL Dataset}
\format{
An object of class \code{list} of length 2.
}
\source{
\href{https://waikato.github.io/weka-wiki/datasets/}{Weka dataset}
(datasets-arie_ben_david.tar.gz, 11,348 Bytes)
}
\usage{
data(esl)
}
\description{
The original ESL data set contains 488 profiles of applicants for certain industrial
jobs.  Expert psychologists of a recruiting company, based upon psychometric
test results and interviews with the candidates, determined the values of the
input attributes. The output is the an overall score corresponding to the
degree of fitness of the candidate to this type of job.
Here classes 1-3 are merged to one group and all classes >6 are merged to one group.
Furtheremore, the data is relabeled to start at 0 and only 20 samples per class are
included.
}
\examples{
data(esl)

}
\references{
ordinal, real-world datasets donated by Dr. Arie Ben David (Holon Inst. of Technology/Israel)
Donor: Arie Ben David
MIS, Dept. of Technology Management
Holon Academic Inst. of Technology
52 Golomb St.
Holon 58102
Israel
abendav@hait.ac.il
Owner: Yoav Ganzah
Business Administration School
Tel Aviv Univerity
Ramat Aviv 69978
Israel
}
\keyword{datasets}
