% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esl_org-data.R
\docType{data}
\name{esl_org}
\alias{esl_org}
\title{ESL Dataset}
\format{
An object of class \code{data.frame} with 488 rows and 5 columns.
}
\source{
\href{https://waikato.github.io/weka-wiki/datasets/}{Weka dataset}
(datasets-arie_ben_david.tar.gz, 11,348 Bytes)
}
\usage{
data(esl_org)
}
\description{
The ESL data set contains 488 profiles of applicants for certain industrial
jobs.  Expert psychologists of a recruiting company, based upon psychometric
test results and interviews with the candidates, determined the values of the
input attributes. The output is the an overall score corresponding to the
degree of fitness of the candidate to this type of job.
}
\examples{
data(esl_org)

}
\references{
ordinal, real-world datasets donated by Dr. Arie Ben David (Holon Inst. of Technology/Israel)
 Donor: Arie Ben David
          MIS, Dept. of Technology Management 
          Holon Academic Inst. of Technology
          52 Golomb St.
          Holon 58102
          Israel
          abendav@hait.ac.il
  Owner: Yoav Ganzah
          Business Administration School
          Tel Aviv Univerity
          Ramat Aviv 69978
          Israel
}
\keyword{datasets}
