% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGraphs.R
\name{generateGraphs}
\alias{generateGraphs}
\title{Generate a dot string graph for the obtained cascades/subcascades}
\usage{
generateGraphs(
  subcascade = NULL,
  useOldLabels = F,
  numericLabels = NULL,
  simplify_iter = 0,
  categoricalLabels = NULL,
  graphName = NULL
)
}
\arguments{
\item{subcascade}{Subcascades object as returned by \code{\link{subcascades}}-function.}

\item{useOldLabels}{Logical; whether to map numeric labels to categorical labels.}

\item{numericLabels}{Optional numeric vector of class labels.}

\item{simplify_iter}{Integer; number of simplification iterations.}

\item{categoricalLabels}{Optional factor or character vector of class labels.}

\item{graphName}{Optional string for graph name}
}
\description{
Generate a dot string graph for the obtained cascades/subcascades
}
