% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{getReport}
\alias{getReport}
\title{Get data}
\usage{
getReport(column = c("impression", "click"), group.by = c("datetime",
  "campaign"), action = "advertiser", filter = NULL, start,
  end = Sys.Date())
}
\arguments{
\item{column}{Variables to retrieve, see details for valid values. Defaults 
to \code{c("impression", "click")}. Must pass  at least 2 values 
(vector of length >= 2). Required.}

\item{group.by}{How to breakdown the data, see @details for valid values. 
Defaults to \code{c("datetime", "campaign")}. Must pass at least 2 values 
(vector of length >= 2). 
Required.}

\item{action}{target node, currently only supports \code{"advertiser"} 
(default).}

\item{filter}{Defaults to \code{NULL}. Optional.}

\item{start}{Start date (YYYY-MM-DD). Required.}

\item{end}{End date (YYYY-MM-DD), defaults to \code{Sys.Date()}. Required.}
}
\description{
Fetch data.
}
\details{
The official documentation can be found 
\href{http://api.ori.cmcm.com/doc/#api-Report-Report_API}{here}.

Valid values for \code{column}:
\itemize{
\item \code{impression}
\item \code{click}
\item \code{conversion}
\item \code{revenue}
\item \code{ctr}
\item \code{cpm}
}

Valid values for \code{group.by}:
\itemize{
\item \code{age}
\item \code{gender}
\item \code{location}
\item \code{brand}
\item \code{ad}
\item \code{adset}
\item \code{campaign}
\item \code{videotype}
}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000, 
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# get daily campaign impressions and conversions for the past 7 days
dat <- getReport(column = c("impression", "conversion"), 
                 group.by = c("datetime", "campaign"), 
                 start = Sys.Date() - 7)
}

}
\seealso{
\code{\link{orionOAuth}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
