% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_adsets.R
\name{listAdsets}
\alias{listAdsets}
\title{List adsets}
\usage{
listAdsets(n = 50)
}
\arguments{
\item{n}{Number of adsets to retrieve, defaults to \code{50}}
}
\description{
Fetch the list of adsets under the authenticated account.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
(adsets <- listAdsets(n = 50))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}
}

