% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_campaigns.R
\name{listCampaigns}
\alias{listCampaigns}
\title{listCampaigns}
\usage{
listCampaigns(n = 50)
}
\arguments{
\item{n}{Number of campaigns to retrieve, defaults to \code{50}}
}
\description{
Fetch the list of campaigns under the authenticated account.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
(camps <- listCampaigns(n = 100))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

