% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_images.R
\name{listImages}
\alias{listImages}
\title{List images}
\usage{
listImages(n = 50)
}
\arguments{
\item{n}{Number of images to retrieve, defaults to \code{50}}
}
\description{
List of images under the authenticated account.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
(images <- listImages())
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}
}

