% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Visualise your account}
\usage{
map(campaigns, adsets, ads, audiences)
}
\arguments{
\item{campaigns}{Data frame of campaigns as returned by 
\code{\link{listCampaigns}}}

\item{adsets}{Data frame of campaigns as returned by 
\code{\link{listAdsets}}}

\item{ads}{Data frame of campaigns as returned by 
\code{\link{listAds}}}

\item{audiences}{Data frame of campaigns as returned by 
\code{\link{listAudiences}}}
}
\description{
Return edge list of objects in the authenticated user's account, 
to help one visualise activities under the account.
}
\details{
Maps the network of object under the authenticated user, follows 
typical structure.

\itemize{
\item{1}{Campaigns}
\item{2}{Adsets}
\item{3}{Ads}
\item{3}{Audiences}
}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list required objects
ads <- listAds(n = 500)
adsets <- listAdsets(n = 500)
campaigns <- listCampaigns(n = 500)
audiences <- listAudiences(n = 500)

network <- map(campaigns, adsets, ads, audiences)

# plot with igraph
# plot ids
g <- igraph::graph.data.frame(network[,1:2], directed = TRUE)
plot(g)

# plot names
g <- igraph::graph.data.frame(network[,3:4], directed = TRUE)
plot(g)

# plot with networkd3
networkD3::simpleNetwork(Data = network, Source = "source.name", 
                         Target = "target.name")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{listAudiences}}, 
\code{\link{listCampaigns}}, \code{\link{listAdsets}}, 
\code{\link{listAds}}
}

