% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonOrdinalData.R
\docType{data}
\name{nonOrdinalData}
\alias{nonOrdinalData}
\title{non-ordinal data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(nonOrdinalData)
}
\description{
It comprises a list of two elements: A data matrix ("data") and 
a label vector ("labels"). The samples are given in the rows
and the features in the columns of the matrix.
The sampled dataset comprises 5 classes with 40 samples, each. 
It is embedded in a two dimensional feature space 
and contains 5 classes with no ordinal sequence.
}
\examples{
data(nonOrdinalData)

}
\keyword{datasets}
