% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalData.R
\docType{data}
\name{ordinalData}
\alias{ordinalData}
\title{ordinal data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(ordinalData)
}
\description{
It comprises a list of two elements: A data matrix ("data") and 
a label vector ("labels"). The samples are given in the rows
and the features in the columns of the matrix.
The sampled dataset comprises 5 classes with 40 samples, each. 
It is embedded in a two dimensional feature space 
and contains 1 ordinal sequence with 5 classes.
}
\examples{
data(ordinalData)

}
\keyword{datasets}
