% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orion_auth.R
\name{orionOAuth}
\alias{orionOAuth}
\title{orionAuthenticate}
\usage{
orionOAuth(client.id, client.secret, grant.type = "client_credentials")
}
\arguments{
\item{client.id}{Your \code{client_id}}

\item{client.secret}{Your \code{client_secret}}

\item{grant.type}{Currently only supports \code{client_credentials} 
(default).}
}
\description{
Authenticate the Orion API
}
\details{
Please see the official documentation to apply for the API and 
get your \code{client_id} and \code{client_secret}: 
\url{api.ori.cmcm.com/doc/#api-Auth-access_token}.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

