% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProjection1D.R
\name{projectData}
\alias{projectData}
\title{1D data projection}
\usage{
projectData(dataset = NULL, comb = NULL)
}
\arguments{
\item{dataset}{list containing at least:
\describe{
\item{data}{Numeric matrix (features x samples)}
\item{labs}{Numeric vector of sample labels}
\item{name}{(optional): dataset name}
\item{oldLabs}{(optional): original categorical sample labels}
}}

\item{comb}{Numeric vector of length 2 specifying the classes for the data projection}
}
\value{
a list with components:
\describe{
\item{data}{one-row-matrix consisting in the 1D mapped datapoints, columns describe each sample}
\item{labs}{Numeric array describing the labels of the samples}
\item{oldLabs}{Categorical array of labels}
}
}
\description{
Generation of 1-dimensional data mapping based on a selected pair of classes
}
