% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeningData.R
\docType{data}
\name{screeningData}
\alias{screeningData}
\title{screening data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(screeningData)
}
\description{
It comprises a list of two elements: A data matrix ("data") and 
a label vector ("labels"). The samples are given in the rows
and the features in the columns of the matrix.
The sampled dataset comprises 6 classes with 40 samples, each. 
It is embedded in a two dimensional feature space 
and contains 2 ordinal sequences with 5 classes.
}
\examples{
data(screeningData)

}
\keyword{datasets}
