% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_adset.R
\name{showAdset}
\alias{showAdset}
\title{Show an adset settings}
\usage{
showAdset(adset.id)
}
\arguments{
\item{adset.id}{id of adset to retrieve}
}
\description{
Retrieve settings of a specfic adset.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list adsets
adsets <- listAdsets(n = 100)

# fetch random adset settings
(adset <- showAdset(adset.id = sample(adsets$id, 1)))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{listAdsets}}
}

