% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_campaign.R
\name{showCampaign}
\alias{showCampaign}
\title{Show a campaign settings}
\usage{
showCampaign(campaign.id)
}
\arguments{
\item{campaign.id}{id of campaign to retrieve}
}
\description{
Retrieve settings of a specfic campaign.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list campaigns
camps <- listCampaigns(n = 100)

# fetch random campaign settings
(camp <- showCampaign(campaign.id = sample(camps$id, 1)))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{listCampaigns}}
}

