% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_object.R
\name{showObject}
\alias{showObject}
\title{Show any object settings}
\usage{
showObject(object, id)
}
\arguments{
\item{object}{Object to retrieve, see details for valid values.}

\item{id}{id of object to retrieve}
}
\description{
Retrieve settings of an object, see details.
}
\details{
Valid values for \code{object}: 
\itemize{
\item \code{audience} see \code{\link{showAudience}} for details.
\item \code{campaign} see \code{\link{showCampaign}} for details.
\item \code{adset} see \code{\link{showAdset}} for details.
\item \code{ad} see \code{\link{showAd}} for details.
}

\code{showObject} can essentially replace any other \code{show} 
family functions. See examples.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list ads
ads <- listObjects(object = "ad", n = 100)

# fetch first ad
ad <- showObject(object = "ad", id = ads$id[1])

# list campaigns
campaigns <- listObjects(object = "campaign")

# fetch random campaign
campaign <- showObject(id = campaigns$id[1])

# equivalent to 
campaign2 <- showCampaign(campaign.id = campaigns$id[1])

identical(campaign, campaign2)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{showAudience}}, 
\code{\link{showCampaign}}, \code{\link{showAdset}} and 
\code{\link{showAd}}
}

