% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_image.R
\name{createImage}
\alias{createImage}
\title{Upload an image}
\usage{
createImage(file)
}
\arguments{
\item{file}{Image data, supports JPEG, GIF and PNG.}
}
\description{
Upload an image
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")

# upload image            
createImage(file = paste0(.libPaths(), "/png/img/Rlogo.png"))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}
}

