% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_campaign.R
\name{deleteCampaign}
\alias{deleteCampaign}
\title{Delete a campaign}
\usage{
deleteCampaign(campaign.id)
}
\arguments{
\item{campaign.id}{id of object to delete}
}
\description{
Delete a campaign.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# set body of request
camp <- list(name = "My Campaign",
             budget_type = "daily",
             budget_daily = "1000", 
             pkg_name = "someting.test.com",
             objective = objective("installs"), 
             app_type = "2",
             web_type = "1")

# create campaign            
campaign <- createCampaign(body = camp)

# delete random campaign
deleteCampaign(campaign.id = campaign$id)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createCampaign}}
}

