% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_object.R
\name{deleteObject}
\alias{deleteObject}
\title{Delete any object}
\usage{
deleteObject(object, id)
}
\arguments{
\item{object}{Object to be deleted, see details for valid values.}

\item{id}{id of object to delete}
}
\description{
Delete campaign, adsets, ads or audience temapltes, 
see details and examples.
}
\details{
Valid values for \code{object}: 
\itemize{
\item \code{audience}, see \code{\link{deleteAudience}} for details.
\item \code{campaign}, see \code{\link{deleteCampaign}} for details.
\item \code{adset}, see \code{\link{deleteAdset}} for details.
\item \code{ad}, see \code{\link{deleteAd}} for details.
}

\code{deleteObject} can essentially replace any other \code{delete} 
family functions. See examples.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list ads
ads <- listObjects(object = "ad")

# delete
# set seed for reproducibility
set.seed(19880525)

# delete random ad
deleteObject(object = "ad", id = sample(ads$id, 1))

# the above is equivalent to
# deleteAd(ad.id = sample(ads$id, 1))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{deleteAudience}}, 
\code{\link{deleteCampaign}}, \code{\link{deleteAdset}} and 
\code{\link{deleteAd}}
}

