% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_object.R
\name{deleteObject}
\alias{deleteObject}
\title{deleteObject}
\usage{
deleteObject(object = "campaign", id)
}
\arguments{
\item{object}{Object to be created, see details for valid values.}

\item{id}{id of object to delete}
}
\description{
Delete campaign, adsets, ads or audience temapltes, 
see details and examples.
}
\details{
Valid values for \code{object}: 
\itemize{
\item \code{audience} see \code{\link{createAudience}} for details.
\item \code{campaign} see \code{\link{createCampaign}} for details.
\item \code{adset} see \code{\link{createAdset}} for details.
\item \code{ad} see \code{\link{createAd}} for details.
}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# get ads
ads <- listObjects(object = "ad")

# delete random ad
deleteObject(object = "ad", id = sample(ads$id, 1))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

