% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_behaviour.R
\name{dictBehaviour}
\alias{dictBehaviour}
\title{List available behaviours}
\usage{
dictBehaviour()
}
\description{
Fetch list of behaviours.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict behaviours
(behaviours <- dictBehaviour())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

