% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_city.R
\name{dictCity}
\alias{dictCity}
\title{dictCity}
\usage{
dictCity()
}
\description{
fetch list of cities
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict cities
(cities <- dictCity())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

