% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_country.R
\name{dictCountry}
\alias{dictCountry}
\title{List valid countries}
\usage{
dictCountry()
}
\description{
Fetch list of countries.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict countries
head(dictCountry())
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

