% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_device.R
\name{dictDevice}
\alias{dictDevice}
\title{List valid device models}
\usage{
dictDevice()
}
\description{
Fetch list of device models.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict mobile devices
(devices <- dictDevice())

# create audience with device setting
createAudience(body = list(
               audience_template_name = "target specific device users",
               device_brand = sample(devices$id, 1)
               ))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createAudience}}
}

