% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_interest.R
\name{dictInterest}
\alias{dictInterest}
\title{List available interests}
\usage{
dictInterest()
}
\description{
Fetch list of interests.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict interests
head(int <- dictInterest())

# filter transportation-related interests
trans <- int[grep("Transportation", int$value),]

# create target audience template 
createAudience(body = list(
               audience_template_name = "target specific interests",
               interest = trans$pid[1]
               ))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}

