% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_language.R
\name{dictLanguage}
\alias{dictLanguage}
\title{List valid languages}
\usage{
dictLanguage()
}
\description{
Fetch list of languages.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict languages
head(lang <- dictLanguage())

# filter russian language code
russian <- subset(lang, value == "Russian")$code

# create audience
createAudience(body= list(
               audience_template_name = "target Russian speakers",
               language = russian
))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createAudience}}
}

