% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_os.R
\name{dictOS}
\alias{dictOS}
\title{List valid OS versions}
\usage{
dictOS()
}
\description{
Fetch list of OS versions.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict os versions
versions <- dictOS()

# create audience
createAudience(body = list(
               audience_template_name = "target android >= 3.0",
               min_device_os = versions[versions$value == "3.0", ]
               ))
}

}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createAudience}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
