% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_state.R
\name{dictState}
\alias{dictState}
\title{List valid states}
\usage{
dictState(country.code)
}
\arguments{
\item{country.code}{Code of country as returned by \code{\link{dictCountry}}, 
see examples.}
}
\description{
Fetch list of states.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# dict countries
country <- dictCountry()

# dict states of random country
states <- dictState(country.code = sample(country$code, 1))

# dict US states
us <- dictState(country.code = "US")

# filter states starting with A (because why not?)
a_states <- us[grep("^A", us$value),]

# create audience
createAudience(body = list(
               audience_template_name = "A-states",
               country = "US",
               region = paste0(a_states$code, collapse="|")
))
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

