% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{age}
\alias{appShowType}
\alias{appType}
\alias{bidType}
\alias{budgetType}
\alias{buttonText}
\alias{deliveryType}
\alias{gender}
\alias{helpers}
\alias{landingPage}
\alias{netType}
\alias{objective}
\alias{switchIt}
\alias{webType}
\title{Get the settings right}
\usage{
budgetType(x)

objective(x)

appType(x)

webType(x)

landingPage(x)

bidType(x)

appShowType(x)

deliveryType(x)

gender(x)

age(x)

netType(x)

buttonText(x)

switchIt(x)
}
\arguments{
\item{x}{Setting to convert}
}
\value{
Returns correct value to pass to API.
}
\description{
These functions attempt to facilitate setting up campaigns, adsets, 
ads and audience targeting templates by allowing to pass more readable 
settings
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# settings with helpers
settings  <- list(name = "test",
                  budget_type = budgetType("daily"),
                  budget_daily = "1000", 
                  pkg_name = "test.com",
                  objective = objective("installs"), 
                  app_type = appType("game"),
                  web_type = webType("ecommerce"), 
                  landing_page = landingPage("googlePlay"))
                  
# settings WITHOUT helpers
nohelpers  <- list(name = "test",
                   budget_type = "daily",
                   budget_daily = "1000", 
                   pkg_name = "test.com",
                   objective = "1", 
                   app_type = "1",
                   web_type = "2", 
                   landing_page = "1")

identical(settings, nohelpers)  
                 
# create campaign
createCampaign(body = settings)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{createCampaign}}, 
\code{\link{createAdset}}, \code{\link{createAd}} and 
\code{\link{createAudience}}
}

