% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_campaigns.R
\name{listCampaigns}
\alias{listCampaigns}
\title{List campaigns}
\usage{
listCampaigns(n = 50)
}
\arguments{
\item{n}{Number of campaigns to retrieve, defaults to \code{50}}
}
\description{
List of campaigns under the authenticated account.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
head(camps <- listCampaigns(n = 100))
}

}
\seealso{
\code{\link{orionOAuth}}
}
\author{
John Coene \email{jcoenep@gmail.com}
}
