% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_objects.R
\name{listObjects}
\alias{listObjects}
\title{listObjects}
\usage{
listObjects(object, n = 50)
}
\arguments{
\item{object}{Nodes to target. See details}

\item{n}{Number of objects to retrieve, defaults to \code{50}}
}
\description{
List any objects from the API, see details for valid 
values.
}
\details{
Valid values for \code{objects}:
\itemize{
\item \code{audience}
\item \code{campaign}
\item \code{adset}
\item \code{ad}
}

\code{listObject} can essentially replace any other \code{list} 
family functions, namely \code{\link{listCampaigns}}. See examples.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list campaigns with listCampaigns
camps <- listCampaigns()

# list campaigns with listObjects
obj <- listObjects(object = "campaign")

# identical results
identical(camps, obj)
}

}
\author{
John Coene \email{john.coene@cmcm.com}
}

