% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ORION-package}
\alias{ORION}
\alias{ORION-package}
\title{ORION: Example Package}
\description{
Package description goes here.
}
\author{
\strong{Maintainer}: HA Kestler \email{hans.kestler@uni-ulm.de}

Authors:
\itemize{
  \item L Lausser
  \item LM Schaefer
  \item R Szekely
}

}
