% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orion-package.R
\docType{package}
\name{orion-package}
\alias{orion}
\alias{orion-package}
\title{orion: Spatial Searching for Euclid}
\description{
The task of finding points closest to a geometry or contained within a geometry from a huge set of points is an important task in many geometry algorithms. 'orion' provides spatial searching for the Euclid ecosystem through a powerful kd tree implementation that supports both nearest/furthest neighbor and range queries.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
