% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oRion.R
\docType{package}
\name{oRion}
\alias{oRion}
\alias{oRion-package}
\title{\pkg{oRion} visualise Twitter Interactions.}
\description{
Interact with Cheetah Mobile's Orion Platform from your R console.
}
\details{
Functions are split in different families.

OAuth
\itemize{
\item \code{\link{orionOAuth}}
}

Get
\itemize{
\item \code{\link{getReport}}
}

\code{create} family
Create items
\itemize{
\item \code{\link{createAd}}
\item \code{\link{createAdset}}
\item \code{\link{createCampaign}}
\item \code{\link{createImage}}
\item \code{\link{createObject}}
}

\code{list} family
List items
\itemize{
\item \code{\link{listAds}}
\item \code{\link{listAdsets}}
\item \code{\link{listCampaigns}}
\item \code{\link{listImages}}
\item \code{\link{listObjects}}
}
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")

# set body of campaign
camp <- list(name = "test",
             budget_type = "daily",
             budget_daily = "1000", 
             pkg_name = "test",
             objective = "1", 
             app_type = "2",
             web_type = "1", 
             landing_page = "2")
             
# create campaign
createCampaign(camp)

# list campaigns
campaigns <- listCampaigns(n = 100)
                      
# create audience template to use in adset                        
createAudience(body = list(
   audience_template_name = "test",
   audience_template_desc = "test",
   country = "CN|BE|IN", 
   language = "fr",
   net_type = 1))

# list audience template
audiences <- listAudiences(n = 50)

# create adset
createAdset(body = list(
   name = paste0("testAdset", i),
   bid_type = 1,
   unit_price = 1000,
   budget_lifetime = 10000,
   campaign_id = sample(campaigns$id, 1),
   audience_create_way = 2,
   app_show_type = 50000,
   audience_template_id = sample(audiences$id, 1),
   click_url = "http://app.adjust.io"))

# list adsets
adsets <- listAdsets()

# upload image to use in ad
createImage(file = "path_to_R_version/library/pn/img/Rlogo.png")

# list images
images <- listImages()

# define adset
body <- list(
   adset_id = adsets$id[1],
   name = "testAd",
   icon_url = images$thumb[1],
   title = "Download oRion",
   desc = "Download oRion and interact with Orion from your R console",
   button_text = 1,
   image_url = images$local_url[1])
   
# post ad
createAd(body = body)

# get data
# get daily campaign impressions and conversions for the past 7 days
dat <- getReport(column = c("impression", "conversion"), 
                 group.by = c("datetime", "campaign"), 
                 start = Sys.Date() - 7)
                  
# get daily adset ratios for the past 30 days by age and gender
dat <- getReport(column = c("ctr", "cpm", "cpc"), 
                 group.by = c("datetime", "adset", "age", "gender"),
                 start = Sys.Date() - 30)
                 
# get impressions and clicks for January 2016's ads
dat <- getReport(column = c("impression", "click"), 
                 group.by = c("datetime", "ad"), start = "2016-01-01",
                 end = "2016-02-01")
}

}

