% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_audience.R
\name{showAudience}
\alias{showAudience}
\title{Show an audience targeting template}
\usage{
showAudience(audience.id)
}
\arguments{
\item{audience.id}{id of audience template to retrieve}
}
\description{
Retrieve settings of a specfic audience targeting template.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")
           
# list ads
audiences <- listAudiences(n = 100)

# fetch random adset settings
(audience <- showAudiences(audience.id = sample(audiences$id, 1)))
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{listAudiences}}
}

