% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_object.R
\name{updateObject}
\alias{updateObject}
\title{Update any object}
\usage{
updateObject(object, body, id)
}
\arguments{
\item{object}{Object to be updated, see details for valid values.}

\item{body}{Body of request that includes campaigns settings. See examples.}

\item{id}{id of object to update}
}
\description{
Update any object, see details and example.
}
\details{
Valid values for \code{object}: 
\itemize{
\item \code{audience} see \code{\link{updateAudience}} for details.
\item \code{campaign} see \code{\link{updateCampaign}} for details.
\item \code{adset} see \code{\link{updateAdset}} for details.
\item \code{ad} see \code{\link{updateAd}} for details.
}

\code{upodateObject} can essentially replace any other \code{update} 
family functions. See examples.
}
\examples{
\dontrun{
# authenticate
orionOAuth(client.id = 0000,
           client.secret = "0x00000000x00x0x000xxx0000x0xx0")

# list current adsets
adsets <- listAdsets()

# new settings                       
settings <- list(switch = switchIt("on"), name = "New Adset Name")

# update adset
update <- updateObject(object = "adset", adset.id = adsets$id[1], 
                       body = settings)

# equivalent to
# update <- updateAdset(adset.id = adsets$id[1], body = settings)

# fetch updated list
adsets <- listAdsets()
}

}
\author{
John Coene \email{jcoenep@gmail.com}
}
\seealso{
\code{\link{orionOAuth}}, \code{\link{updateAudience}}, 
\code{\link{updateCampaign}}, \code{\link{updateAdset}}, 
\code{\link{updateAd}} and \code{\link{helpers}}
}

